added get and set functions for data and Nodes
Get and set methods set the data which will be cards and returns it when with the get method

Get and set methods for the node can get the node from the [] and you can set the next node for the []

while working on these functions I noticed that someone added a addNext function with an incomplete heading? I added self but unsure about the n : Node so I left it alone
Toni if you dont see this error it says Node is not defined
we got this guys! :smiley: